Create MYSQL 1
LAB ASSIGNMENT 5 (MYSQL) 

Write the SQL commands to create the table MOVIES  with the following constraints 

MovieNo   Primary Key 

Title          UNIQUE 

TYPE        NOT NULL 

                                   
RELATION:MOVIES 
Dislay the details of the movies having  starts as NULL value 

Display the Title and Rating  of the movie type COMEDY 

Display the details of movies having quantity more than 3 

Display the details of Drama movies having price in the range of 50 to 100 

Display the details of movies with stars as “Cruise”, “Harris”, ”Hoffman”. 

Display a list of all movies with Price over 20 and sorted by price. 

Display all the movies sorted by QTY in decreasing order. 

Display a report listing books id, Title and current value(Qty* Price). 

Display the names of movies with title starting with the letter “C” 

Display the names of movies with title  with “o”  as second character. 

 Display the names of movies with title with the length of the title name